Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
alembic
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.12.0
aiofiles

